% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remav2-package.R
\docType{package}
\name{remav2-package}
\alias{remav2}
\alias{remav2-package}
\title{remav2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
