% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remav2_index.R
\name{remav2_index}
\alias{remav2_index}
\title{REMA v2 Mosaic Index}
\usage{
remav2_index()
}
\value{
data frame with mosaic index details
}
\description{
Data frame with relevant tile identifiers, source url, and COG urls for REMA v2 2m tiles.
}
\details{
The data frame \link{remav2_mosaic_index} has the following structure.

\tabular{rlll}{
\tab dem_id   \tab character \tab dem_id\cr
\tab tile   \tab character \tab tile\cr
\tab xmin   \tab numeric \tab xmin\cr
\tab xmax   \tab numeric \tab xmax\cr
\tab ymin   \tab numeric \tab ymin\cr
\tab ymax   \tab numeric \tab ymax\cr
\tab s3url   \tab character \tab s3url\cr
\tab data_percent   \tab numeric \tab data_percent\cr
\tab num_components   \tab numeric \tab num_components\cr
\tab cog_browse   \tab character \tab cog_browse\cr
\tab cog_dem   \tab character \tab cog_dem\cr
\tab cog_count   \tab character \tab cog_count\cr
\tab cog_countmt   \tab character \tab cog_countmt\cr
\tab cog_mad   \tab character \tab cog_mad\cr
\tab cog_maxdate   \tab character \tab cog_maxdate\cr
\tab cog_mindate   \tab character \tab cog_mindate\cr
}
}
\examples{
remav2_index()
}
