% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remav2-package.R
\docType{data}
\name{remav2_mosaic_index}
\alias{remav2_mosaic_index}
\alias{REMA_Mosaic_Index_v2_2m}
\title{The REMA v2 Mosaic Index (2m)}
\description{
The Reference Elevation Model of Antarctica v2, this index is a set of tile identifiers and geographic extents
with links to the s3 STAC ('s3url'), the tarball sources ('fileurl'), and the polygon of the tile and other information.
}
\details{
This is preserved as provided in the mosaic index Geopackage 'REMA_Mosaic_Index_v2_2m', but we also augment this with
the various cog sources (layers provided in the tarballs, also available at these COG urls) and the extent of each in 'remav2_mosaic_index'.

The function \code{\link[=remav2_index]{remav2_index()}} returns the remav2_mosaic_index as a convenience.
}
\seealso{
remav2_index
}
