% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remav2_index.R
\name{remav2_index}
\alias{remav2_index}
\title{REMA v2 Mosaic Index}
\usage{
remav2_index()
}
\description{
Data frame with relevant tile identifiers, source url, and COG urls for REMA v2 2m tiles.
}
\details{

}
